+++ HISTORY.md
@@ -4,6 +4,7 @@
 ## not yet released, version 0.2.0
 
 - Implemented support for publish/subscribe in the peer-to-peer network.
+- Implemented `Host.get(id)` to get an existing Peer by its id.
 - Changed `Host.create(id)` and `Host.remove(id)` to a synchronous calls 
   (instead of returning Promises).
 - Fixed a bug in cleaning up after failing to connect to a non-existing host.